Moved user files inside user folder
