+++ src/App.jsx
@@ -15,10 +15,10 @@ import SigninOptions from './components/pages/onBoarding/SigninOptions'
 import Favourites from './components/pages/user/favourites/Favourites'
 import ForgotPassword from './components/pages/login/ForgotPassword'
 import VerifyCode from './components/pages/login/VerifyCode'
-import AlternativesPage from './components/pages/AlternativesPage'
-import SearchResults from './components/pages/SearchResults'
+import AlternativesPage from './components/pages/user/AlternativesPage'
+import SearchResults from './components/pages/user/SearchResults'
 import Main from './components/pages/user/main/Main'
-import Prescription from './components/pages/Prescription'
+import Prescription from './components/pages/user/Prescription'
 import SignUp from './components/pages/signUp/SignUp'
 import InsuranceCompanies from './components/pages/user/insurance/InsuranceCompanies'
 import CompanyPage from './components/pages/user/insurance/CompanyPage'
